+++ src/dataframe/dataframe.jl
@@ -13,13 +13,14 @@ DataFrame(pairs::AbstractVector{<:Pair}; makeunique::Bool=false, copycols::Bool=
 DataFrame(ds::AbstractDict; copycols::Bool=true)
 DataFrame(; kwargs..., copycols::Bool=true)
 
-DataFrame(columns::AbstractVecOrMat,
-          names::AbstractVector;
+DataFrame(table; copycols::Union{Bool, Nothing}=nothing)
+DataFrame(table, names::AbstractVector;
+          makeunique::Bool=false, copycols::Union{Bool, Nothing}=nothing)
+DataFrame(columns::AbstractVecOrMat, names::AbstractVector;
           makeunique::Bool=false, copycols::Bool=true)
 
-DataFrame(table; copycols::Union{Bool, Nothing}=nothing)
-DataFrame(::DataFrameRow)
-DataFrame(::GroupedDataFrame; keepkeys::Bool=true)
+DataFrame(::DataFrameRow, copycols::Bool=true)
+DataFrame(::GroupedDataFrame; copycols::Bool=true, keepkeys::Bool=true)
 ```
 
 # Keyword arguments
@@ -64,6 +65,11 @@ If a single positional argument is passed to a `DataFrame` constructor then it
 is assumed to be of type that implements the
 [Tables.jl](https://github.com/JuliaData/Tables.jl) interface using which the
 returned `DataFrame` is materialized.
+
+Additionally, if two positional agruments are passed, where the second argument is
+`AbstractVector`, then the first argument is passed to `DataFrame` constructor
+with an appropriate `copycols` argument and later the returned data frame gets new
+column names that are taken from the vector that is a second positional argument.
 
 Finally it is allowed to construct a `DataFrame` from a `DataFrameRow` or a
 `GroupedDataFrame`. In the latter case the `keepkeys` keyword argument specifies
@@ -330,7 +336,7 @@ end
 function DataFrame(columns::AbstractVector, cnames::AbstractVector{Symbol};
                    makeunique::Bool=false, copycols::Bool=true)::DataFrame
     if !(eltype(columns) <: AbstractVector) && !all(col -> isa(col, AbstractVector), columns)
-        throw(ArgumentError("columns argument must be a vector of AbstractVector objects"))
+        return rename!(DataFrame(columns, copycols=copycols), cnames, makeunique=makeunique)
     end
     return DataFrame(collect(AbstractVector, columns),
                      Index(convert(Vector{Symbol}, cnames), makeunique=makeunique),